+++ repos/os/src/drivers/sd_card/rpi/driver.cc
@@ -263,7 +263,7 @@ void Driver::read(Block::sector_t           block_number,
 {
 	_set_block_count(block_count);
 
-	if (!issue_command(Read_multiple_block(_block_to_command_address(block_number)))) {
+	if (!issue_command(Read_multiple_block(_block_to_command_address((size_t)block_number)))) {
 		error("Read_multiple_block failed");
 		throw Io_error();
 	}
@@ -302,7 +302,7 @@ void Driver::write(Block::sector_t           block_number,
 {
 	_set_block_count(block_count);
 
-	if (!issue_command(Write_multiple_block(_block_to_command_address(block_number)))) {
+	if (!issue_command(Write_multiple_block(_block_to_command_address((size_t)block_number)))) {
 		error("Write_multiple_block failed");
 		throw Io_error();
 	}
